feat: refactor type imports to use index.js and enhance code organization
