+++ src/services/claim.api.service.ts
@@ -5,7 +5,8 @@ import {
   MidnightApiError,
   SubmitClaimResponse,
   SupportedBlockchains,
-} from "../types.js";
+} from "../types/index.js";
+
 import axiosInstance from "../utils/httpClient.js";
 import { Logger } from "../utils/logger.js";
 import { buildCoseSign1 } from "../utils/cardano.utils.js";
